// src/models/errors/unauthenticatederror.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v3";
import { nullToUndefined } from "../../lib/schemas.js";
import { LambdaDBError } from "./lambdadberror.js";

export type UnauthenticatedErrorData = {
  message?: string | undefined;
};

export class UnauthenticatedError extends LambdaDBError {
  /** The original data that was passed to this error instance. */
  data$: UnauthenticatedErrorData;

  constructor(
    err: UnauthenticatedErrorData,
    httpMeta: { response: Response; request: Request; body: string },
  ) {
    const message = err.message || `API error occurred: ${JSON.stringify(err)}`;
    super(message, httpMeta);
    this.data$ = err;

    this.name = "UnauthenticatedError";
  }
}

/** @internal */
export const UnauthenticatedError$inboundSchema: z.ZodType<
  UnauthenticatedError,
  z.ZodTypeDef,
  unknown
> = z.object({
  message: nullToUndefined(z.string().optional()),
  request$: z.instanceof(Request),
  response$: z.instanceof(Response),
  body$: z.string(),
})
  .transform((v) => {
    return new UnauthenticatedError(v, {
      request: v.request$,
      response: v.response$,
      body: v.body$,
    });
  });