// src/models/errors/resourcealreadyexistserror.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v3";
import { nullToUndefined } from "../../lib/schemas.js";
import { LambdaDBError } from "./lambdadberror.js";

export type ResourceAlreadyExistsErrorData = {
  message?: string | undefined;
};

export class ResourceAlreadyExistsError extends LambdaDBError {
  /** The original data that was passed to this error instance. */
  data$: ResourceAlreadyExistsErrorData;

  constructor(
    err: ResourceAlreadyExistsErrorData,
    httpMeta: { response: Response; request: Request; body: string },
  ) {
    const message = err.message || `API error occurred: ${JSON.stringify(err)}`;
    super(message, httpMeta);
    this.data$ = err;

    this.name = "ResourceAlreadyExistsError";
  }
}

/** @internal */
export const ResourceAlreadyExistsError$inboundSchema: z.ZodType<
  ResourceAlreadyExistsError,
  z.ZodTypeDef,
  unknown
> = z.object({
  message: nullToUndefined(z.string().optional()),
  request$: z.instanceof(Request),
  response$: z.instanceof(Response),
  body$: z.string(),
})
  .transform((v) => {
    return new ResourceAlreadyExistsError(v, {
      request: v.request$,
      response: v.response$,
      body: v.body$,
    });
  });